fix: update image validation types and upload path
- change max_size type from int to float in both the route and service to accept fractional MB values
- reformat function parameters/calls for readability
- and fix the validation error message to reference expected_extensions.
- also update the static uploads mount from "/uploads" to "/api/uploads" to align with the API namespace.
- changes affect routes.py, main.py, and image_validator.py.

+++ server/app/api/routes.py
@@ -6,11 +6,24 @@ from app.services.image_validator import validate_image_service
 router = APIRouter(prefix="/api", tags=["validation"])
 
 @router.post("/validate", response_model=ValidationResponse)
-async def validate_image(files: List[UploadFile] = File(...), max_size: int | None = Form(None), expected_width: int | None = Form(None), expected_height: int | None = Form(None), expected_extensions: str | None = Form(None)):
+async def validate_image(
+    files: List[UploadFile] = File(...), 
+    max_size: float | None = Form(None), 
+    expected_width: int | None = Form(None), 
+    expected_height: int | None = Form(None), 
+    expected_extensions: str | None = Form(None)
+):
     results = []
 
     for file in files:
-        result = validate_image_service(file, max_size, expected_width, expected_height, expected_extensions)
+        
+        result = validate_image_service(
+            file, 
+            max_size, 
+            expected_width, 
+            expected_height, 
+            expected_extensions
+        )
 
         results.append({
             "filename": file.filename,
